+++ src/main/java/kitchenpos/eatinorder/infra/JpaOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.eatinorder.domain;
+package kitchenpos.eatinorder.infra;
 
+import kitchenpos.eatinorder.domain.Order;
+import kitchenpos.eatinorder.domain.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
