+++ src/main/java/kitchenpos/eatinorder/infra/JpaOrderTableRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.eatinorder.domain;
+package kitchenpos.eatinorder.infra;
 
+import kitchenpos.eatinorder.domain.OrderTable;
+import kitchenpos.eatinorder.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
