+++ src/main/java/kitchenpos/product/domain/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.product.infra;
+package kitchenpos.product.domain;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
